Promote sercom::v2 to sercom
