+++ hal/src/gpio/mod.rs
@@ -16,8 +16,7 @@
 //!
 //! ## Features
 //!
-//! - Converting between pin modes no longer requires access to the [`Port`]
-//!   type.
+//! - Converting between pin modes no longer requires access to the `Port` type.
 //!
 //! As a consequence, pin mode conversions can now be implemented using
 //! [`From`]/[`Into`].
