+++ hal/src/sercom/i2c/flags.rs
