+++ hal/src/sercom/i2c/impl_ehal.rs
