+++ hal/src/sercom/pad.rs
