+++ hal/src/sercom/pad/impl_pad_thumbv6m.rs
@@ -1,7 +1,7 @@
 //! Implementations of the [`IsPad`], [`IsI2cPad`], [`GetPad`] traits
 
 use crate::gpio::*;
-use crate::sercom::v2::*;
+use crate::sercom::*;
 
 //==============================================================================
 //  Pad definitions
