+++ hal/src/sercom/spi/char_size.rs
