+++ hal/src/sercom/spi/impl_ehal_thumbv6m.rs
