+++ hal/src/sercom/spi/impl_ehal_thumbv7em.rs
