+++ hal/src/sercom/spi/pads_thumbv6m.rs
@@ -7,7 +7,7 @@ use core::marker::PhantomData;
 
 #[cfg(feature = "samd21")]
 use crate::gpio::AnyPin;
-use crate::sercom::v2::*;
+use crate::sercom::*;
 use crate::typelevel::{NoneT, Sealed};
 
 use super::{Capability, Duplex, Rx, Tx};
@@ -389,7 +389,7 @@ where
 /// ```
 /// use atsamd_hal::pac::Peripherals;
 /// use atsamd_hal::gpio::{PA08, PA09, Pins};
-/// use atsamd_hal::sercom::v2::{Sercom0, spi};
+/// use atsamd_hal::sercom::{Sercom0, spi};
 /// use atsamd_hal::typelevel::NoneT;
 ///
 /// pub type Pads = spi::PadsFromIds<Sercom0, PA08, NoneT, PA09>;
