+++ hal/src/sercom/uart/charsize.rs
