+++ hal/src/sercom/uart/flags.rs
