+++ hal/src/sercom/uart/impl_ehal.rs
